first support for tangent vertex attribute

+++ src/GlMesh.hpp
@@ -91,7 +91,7 @@ private:
     GlMesh(const GlMesh&);
     GlMesh& operator = (const GlMesh&);
     size_t getBufferIndex(const std::string &name);
-    void generate(const vfm::ObjModel &objModel, unsigned int channel, std::vector<GLfloat> &buffer);
+    void generate(const vfm::ObjModel &objModel, size_t channel, std::vector<GLfloat> &buffer);
     void clear();
 
     GLuint _vertexArray;
